# add storage API to speed up data loading