@m-d-bowerman 2024-03-07
I believe the `to_dataframe` method already does this under the hood, based on the `create_bqstorage_client` [parameter](https://github.com/googleapis/python-bigquery/blob/08762fbf3289622c2b948da919ae49567d0378d5/google/cloud/bigquery/table.py#L2087-L2095) that defaults to true.